pass SkipUpdate in ScanOptions to stop app updates
the `--skip-update` option didn't prevent application scan
updates as those happend a bit later in the whole process.

we've passed those settings through the "scanoptions"
to get that handled later in the system so we can prevent
the UpdateDB call from happening during the scan.

this also updates ScanFile to also accept "ScanOptions"