fix(doc): gitlab integration
The GitLab CI yml alernative does not work.

- yml invalid: `artifacts:paths` is a list and a `-` is missing
- `--template` option: there is no `contrib/gitlab-codeclimate.tpl` file, you meant `contrib/gitlab-codequality.tpl` instead

And I did not suggest a change, but I had to change the docker images from 'docker:stable' and 'docker:dind' respectively to 'docker:19-stable' and 'docker:19-dind'. This is probably due to my gitlab-runner configuration, but a note could help some others in my situation.