Detect packages when --list-all-pkgs provided no matter the vuln type configuration
Resolves #828

The current workflow of the `scan` module is calling the `checkVulnerabilities` function
with is vulnerabilities oriented.
It checks each vulnerability type (currently library and OS) that it was provided with.
In addition, in case `--list-all-pkgs` is specified, it also looks for each
vulnerability type packages.

This PR supports that when providing the `--list-all-pkgs` flag, it will look into
every vulnerability type, no matter the `--vuln-type` flag that was provided.

This behavior should require more design changes but for now, in order to keep this code
backward compatible, this is a valid solution.