Fix v3_vector location and add redhat version
This commit fixes the location on the input data where the NVD v3 vector
format data can be found.  The previous behaviour caused the rule to
always evaluate to false.

The commit also adds the RedHat NVD v3 vector location and extends the
rules to use it as well.  We found examples where the RedHat data was
populated but not the NVD data where this is helpful.