Revise integration tests
An unintended consequence of importing with the option
`containerd.WithIndexName(tt.imageName)` was that it was changing the
digest of the image with each import, because the name of the original
image (ghcr.io/aquasecurity/trivy-test-images:alpine-310) gets included
in the annotations section of the manifest pointed to by index.json.
That of course changes the hash of the manifest to `63775...`.

This pull request ensures that the imported image retains the same
digest by importing normally, then tagging the image with the desired
name. The original tag `ghcr.io/aquasecurity/trivy-test-images:alpine-310`
is immediately removed to ensure the image is only known by the name we
have given it.

Additionally, an index guard has been provided in `pkg/fanal/image/daemon/containerd.go`.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>