Adding --module-dir and --enable-modules
Adding a global flag which accepts an input dir where WASM files are to be loaded from

Passing in the input string of dir to be loaded / registered as WASM modules from.

Adding --enable-modules flag which If nothing is specified, it loads all modules from --module-dir/dir(). If --enable-modules is specified, we enable those modules only.