feat: add built-in rule for JWT tokens
Currently the secret scanning does not detect JWT tokens, which are used for example by Artifactory as Authentication mechanism. 

For example, if a user builds an OCI image and installs dependencies from a private Artifactory repository during build-time and has passed those credentials in an insecure manner to the builder (for example using ARG statements), trivy needs to detect this leak.