chore: add an issue template for maintainers
### Description
Due to recent changes on GitHub, maintainers are currently unable to create issues directly. This is inconvenient, so we're temporarily adding a dedicated issue template for maintainers. You can test it [here](https://github.com/aquasecurity/trivy-test/issues/new/choose).

### Caveat
Ideally, the maintainer template would appear at the bottom of the list, but based on my research, it's not possible to place external links at the top. As a result, the maintainer template appears first.