refactor(cli): Rework Trivy Cloud integration logic
After having discussions internally, the logic for working with the
Trivy Cloud integration is going to be simplified greatly. This involves
removing the Cloud specific config file and bringing the config into the
existing Trivy config. The `cloud config` commands will be removed as
they aren't needed anymore.

As the focus is going to be on CI/CD integration with the Trivy Cloud
platform, there isn't currently a need to have the Login and Logout
mechanisms.