feat: add browser login
Add login/logout flow to get a token from the API using the browser.

Login with trivy login, if providing --token, that will be stored in the
keyring, if not, the browser will be opened up to login and generate a
token and callback to the cli before saving to the keyring.

Some small updates to the log prefix

Generate the docs for the new commands